Protect swap internal from 0 swap, fix tick spacing return

+++ pkg/sol/SeawaterAMM.sol
@@ -410,7 +410,7 @@ contract SeawaterAMM is ISeawaterAMM {
     }
 
     /// @inheritdoc ISeawaterExecutorAdminExposed
-    function tickSpacing653FE28F(address /* pool */) external returns (uint16) {
+    function tickSpacing653FE28F(address /* pool */) external returns (uint8) {
         directDelegate(_getExecutorAdmin());
     }
